@@ -3,6 +3,8 @@ package io.github.Vz0n.neko.util;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 
 import javax.annotation.Nullable;
@@ -16,35 +18,36 @@ import org.json.simple.parser.ParseException;
 public class HttpUtil {
 
     @Nullable
-    public static JSONObject getJSONResponse(String url) throws IOException, ParseException {
+    public static JSONObject getJSONResponse(String url) throws IOException, ParseException, URISyntaxException {
 
           JSONParser parser = new JSONParser();
-          HttpsURLConnection htsconn = (HttpsURLConnection) 
-                               new URL(url).openConnection();
+
+          // Create URI and connection
+          URL urlObj = new URI(url).toURL();
+          HttpsURLConnection httpsConn = (HttpsURLConnection) urlObj.openConnection();
 
           // Server is not available
-          if(htsconn.getResponseCode() != 200) return null;
+          if(httpsConn.getResponseCode() != 200) return null;
     
             
-          InputStream stream = htsconn.getInputStream();
-          JSONObject obj = (JSONObject) parser.parse(
-            new String(stream.readAllBytes())
+          InputStream stream = httpsConn.getInputStream();
+
+          return (JSONObject) parser.parse(
+                  new String(stream.readAllBytes())
           );
-               
-          return obj;
           
     }
  
     @Nullable
-    public static BufferedImage getImage(String url) throws IOException {      
-      
-        HttpsURLConnection conn = (HttpsURLConnection) 
-                               new URL(url).openConnection();
+    public static BufferedImage getImage(String url) throws IOException, URISyntaxException {
+
+        URL urlObj = new URI(url).toURL();
+        HttpsURLConnection httpsConn = (HttpsURLConnection) urlObj.openConnection();
 
         // Response isn't a image and/or server is unavailable.
-        if(!conn.getContentType().startsWith("image/")) return null;
+        if(!httpsConn.getContentType().startsWith("image/")) return null;
 
-        return ImageIO.read(conn.getInputStream());
+        return ImageIO.read(httpsConn.getInputStream());
 
     }
 
